# tribute-api/static/openapi.yaml
openapi: 3.0.0
info:
  title: Tribute Image Attribution API
  description: REST API to process images and extract referenced artists
  version: 1.0.0
servers:
  - url: https://api.art-tribute.org
paths:
  /process_image_url:
    get:
      operationId: process_image_url
      summary: Process an image from the provided URL
      parameters:
        - name: image_url
          in: query
          description: URL of the image to be processed
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully processed the image and extracted referenced artists
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      artists:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                      possible_prompt:
                        type: string
                      possible_sources:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                            data:
                              type: object
                              properties:
                                source:
                                  type: string
                                chain:
                                  type: string
                                offchain_address:
                                  type: string
                                onchain_address:
                                  type: string
                      similar_images:
                        type: array
                        items:
                          type: object
                          properties:
                            image_url:
                              type: string
        '400':
          description: Missing image URL in the request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Error processing the image or no output data from Huggingface API
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string